requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=10.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
werkzeug>=3.0.0

